Split psychometric export responsibilities
