+++ apps/backend/src/app/coding/coding.module.ts
@@ -45,7 +45,11 @@ import {
   CodingFreshnessService,
   CodingReadinessService,
   CodingItemMatrixExportService,
-  CodingPsychometricExportService
+  CodingPsychometricExportService,
+  PsychometricMetadataResolver,
+  PsychometricResponseReader,
+  PsychometricAnalysisEngine,
+  PsychometricExportWriter
 } from '../database/services/coding';
 import { CODING_PROCESS_CACHE_INVALIDATOR } from '../database/services/coding/coding-process-cache-invalidator.token';
 import { CODING_READINESS_CACHE_INVALIDATOR } from '../database/services/coding/coding-readiness-cache-invalidator.token';
@@ -112,6 +116,10 @@ import { UserModule } from '../user/user.module';
     CodingFreshnessService,
     CodingReadinessService,
     CodingItemMatrixExportService,
+    PsychometricMetadataResolver,
+    PsychometricResponseReader,
+    PsychometricAnalysisEngine,
+    PsychometricExportWriter,
     CodingPsychometricExportService,
     {
       provide: CODING_PROCESS_CACHE_INVALIDATOR,
@@ -150,4 +158,4 @@ import { UserModule } from '../user/user.module';
     CODING_READINESS_CACHE_INVALIDATOR
   ]
 })
-export class CodingModule { }
+export class CodingModule {}
